requirements.txt
chevron==0.14.0
jsonschema==4.23.0
openai>=1.0.0
python-dotenv>=1.0.0

